[FEATURE] track picked changes in beard.lock
This commit adds a new file beard.lock when it picks
changes to track the commitHash of the local commit
in relation to the original commitHash of the remote
revision. previously this was done by "littering" the commit
message with lines like these: "(cherry picked from commit...)"
this is now a thing of the past.